move make token to token-utils
