+++ packages/stableswap-sdk/package.json
@@ -10,7 +10,7 @@
     "url": "https://github.com/saber-hq/saber-common.git"
   },
   "author": "Saber Team <team@saber.so>",
-  "version": "1.12.9",
+  "version": "1.12.10",
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
   "license": "Apache-2.0",
@@ -37,7 +37,7 @@
   },
   "dependencies": {
     "@saberhq/solana-contrib": "^1.12.9",
-    "@saberhq/token-utils": "^1.12.9",
+    "@saberhq/token-utils": "^1.12.10",
     "@solana/buffer-layout": "^4.0.0",
     "lodash.mapvalues": "^4.6.0",
     "tslib": "^2.3.1"
